Converted CountAggregator to work with doubles so that it is easier to generalize with its peers

+++ src/main/scala/com/uncharted/mosaic/generation/analytic/numeric/CountAggregator.scala
@@ -2,20 +2,20 @@ package com.uncharted.mosiac.generation.analytic.numeric
 
 import com.uncharted.mosiac.generation.analytic.Aggregator
 
-object CountAggregator extends Aggregator[Any, Long, java.lang.Double] {
+object CountAggregator extends Aggregator[Any, Double, java.lang.Double] {
 
-  def default(): Long = {
-    0L
+  def default(): Double = {
+    0D
   }
 
-  override def add(current: Long, next: Option[Any]): Long = {
+  override def add(current: Double, next: Option[Any]): Double = {
     current + 1
   }
-  override def merge(left: Long, right: Long): Long = {
+  override def merge(left: Double, right: Double): Double = {
     left+right
   }
 
-  def finish(intermediate: Long): java.lang.Double = {
+  def finish(intermediate: Double): java.lang.Double = {
     intermediate.toDouble
   }
 }
